+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v1.0 (in development, master branch)
 - Dropped support for Python 2
 
 ### Bot changes
+- ENH: added bots.collectors.rt.collector
 - ENH: added bots.parsers.spamhaus.parser_cert
 - ENH: added bots.parsers.fraunhofer.parser_dga
 - ENH: added bots.experts.certat_contact.expert